requirements.txt
pyobjc
# pyobjc-core==11.0
# pyobjc-framework-Cocoa==11.0
# pyobjc-framework-Quartz
# pyobjc-framework-UserNotifications
# pyobjc-framework-AppKit
py2app
